fix: Removed duplicate files, renamed components
