+++ src/assets/icons/index.js
@@ -1,8 +1,4 @@
-export { default as space } from './space.jpg';
-export { default as lens } from './magnify.png';
-// export { default as gear } from './magnify.png';
-export { default as japan } from './japan.jpg';
-export { default as arrow } from './arrow.png';
+
 export { default as banner } from './banner.png';
 export { default as tmpBanner } from './banner2.jpg';
 export { default as list } from './list.png';
@@ -10,10 +6,8 @@ export { default as upwardArrow } from './upwards-arrow.png';
 export { default as gear } from './gears.svg';
 export { default as logoWithTxt } from './logo_text_png.png';
 export { default as cancel } from './cancel.svg';
-export { default as tmpBg } from './engine.jpg';
 export { default as rightArrow } from './right-arrow.svg';
 export { default as profile } from './avatar.svg';
-// export { default as banner } from './banner.jpeg';
 export { default as category_ic } from './copy.svg';
 export { default as kids } from './kids.png';
 export { default as whyUs } from './whyUs.jpg';
@@ -25,5 +19,5 @@ export { default as decide } from './decide.jpg';
 export { default as explore } from './explore.jpg';
 export { default as logo } from './logo_png.png';
 export { default as school } from './school.png';
-export { default as HomePgBanner } from './Banner_2.png';
+export { default as HomePgBanner } from './HomePgBanner.png';
 export { default as category } from './category.png';
