+++ src/components/cardWithDetailsBelow/index.js
