+++ src/components/categoryLabel/index.js
