+++ src/components/headerForBlog/animationStyle.css
