+++ src/components/socialIcons/index.js
@@ -7,7 +7,6 @@ import {
   faUsers
 } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import Tag from '../tags_b';
 
 const SocialIcons = () => (
   <div className="social-icons">
