+++ src/components/socialIcons/style.scss
