--- src/constants/style.js
@@ -1,71 +1,0 @@
-export default {
-  // Colors Palette
-  blue: {
-    cobalt: '#004D9C',
-    mayaBlue: '#63BCFD',
-    sky: '#0066C0',
-    navyBlue: '#0066C0',
-    darkBlue: '#054484'
-  },
-  black: {
-    black: '#333333',
-    dimGray: '#666666',
-    lightGray: '#444444',
-    nobel: '#999999',
-    silver: '#C4C4C4',
-    gainsboro: '#DFDFDF'
-  },
-  white: {
-    white: '#ffffff',
-    whiteSmoke: '#F2F2F2',
-    snow: '#FFEFEF'
-  },
-  red: {
-    speechRed: '#A82205',
-    alizarin: '#E01E1E',
-    orange: '#FF4E00'
-  },
-  green: {
-    green: '#079A18',
-    latte: '#ECFCEE',
-    olive: '#6F6A00',
-    chiffon: '#F2EFAF'
-  },
-  yellow: {
-    tulip: '#E6BF18',
-    sand: '#F4D384'
-  },
-
-  // spacings
-  screenPadding: '20px 15px',
-  xs: '5px',
-  sm: '10px',
-  md: '15px',
-  lg: '20px',
-  xl: '25px',
-  xxl: '30px',
-
-  // font sizes
-  fontSize: {
-    fs10: '10px',
-    fs11: '11px',
-    fs12: '12px',
-    fs13: '13px',
-    fs14: '14px',
-    fs16: '16px',
-    fs18: '18px',
-    fs20: '20px',
-    fs22: '22px',
-    fs24: '24px',
-    fs26: '26px',
-    fs28: '28px',
-    fs30: '30px',
-    headerText: '16px'
-  },
-  fontWeight: {
-    regular: 400,
-    semiBold: 600,
-    bold: 700
-  },
-  fontFamily: 'Open Sans'
-};
