+++ src/containers/blogPage/style.scss
@@ -8,7 +8,7 @@ $white:#ffffff;
     width:60%;
     margin:auto;
     // background-color: pink;
-    padding-bottom: 15px;
+   padding-bottom: 15px;
 
 }
 
@@ -152,6 +152,7 @@ $white:#ffffff;
   .blogContainer{
     width:90%;
     margin:auto;
+     padding: 10% 15px;
   }
    .recommendedTitle{
         font-size: 30px;
